Moved App.svelte into a components folder.
