+++ webapp/src/main.js
@@ -1,11 +1,8 @@
 /* eslint-disable no-undef */
-import App from './App.svelte';
+import App from './components/App.svelte';
 
 const app = new App({
-  target: document.body,
-  props: {
-    name: 'world'
-  }
+  target: document.body
 });
 
 export default app;
